Moved PermissionsProvider to permissions module
