+++ permissions/src/main/res/drawable/ic_get_accounts.xml
