+++ permissions/src/main/res/drawable/ic_mic.xml
