+++ permissions/src/main/res/drawable/ic_phone.xml
