+++ permissions/src/main/res/drawable/ic_photo_camera.xml
