+++ permissions/src/main/res/drawable/sd.xml
